finally found a way to split the component template html out AND still compile as a UMD
bundling up a component with a relative, or a verbose, file path to an HTML asset would break after bundling since the parent module would just look for the asset based on its own path (“./foo.html” would 404 in the parent module since it only exists in the child module) and for whatever reason using module.id just doesn’t work as expected anymore.

i tried using a rollup plugin to consume and minify the file itself, but it seems to just ignore the path of `templateUrl` despite what the docs say.

I finally moved all template html into a ‘fake’ typescript module that just exports a var called ‘tpl’. we can use this var as the string we use in our angular component which both compiles, is able to be tree-shaken and allows us to bundle up a package. i’d like to note this HTML is still minified.